Add reply_target concept.
The `reply_target` of an event is the sender's nickname when we are the
`target`, otherwise, it is the same as `target`.

It lets us reply to events without having to repeat the `(me ===
target) ? sender.nick : target` logic used to make sure we don't message
ourselves when we receive a direct PRIVMSG, NOTICE, etc.

For ease, Bot::AmbientVerbs now default to targeting the ambient
`reply_target`. (Aliased as `ambient_reply_target`).