Add reply target concept
The `reply_target` is the sender's nickname when we are the `target`,
otherwise, it's equal to `target`.

It lets us reply to events without having to repeat the logic used to
make sure we don't message ourselves when we receive a direct PRIVMSG,
NOTICE, etc.

Bot::AmbientVerbs now target the ambient `reply_target` by default.